lower the ISR stack size by default
STM32 is very tight on ram. Enabling even a few things like combos and holdtaps can cause the memory needs to grow larger than what is available.

This is currently left to the end user to troubleshoot, but I think if we set a sensible default here it should stave off most issues.

There are other knobs that can be tuned, and I may add those later if I run into issues myself or if I figure them out enough to be confident setting a lower default value.